WIP: rename route-rule to routerule
This contains a dirty regex hack. Test cases updated to use real model
types, but I wonder if the logic here is sufficiently future-proof.

I don't see another way to do it as the original logic expects input in
the format `route-rule-x` to output `RouteRuleX`. With the first hyphen
removed (`routerule-x`) there is no easy way to use the previous logic
without either iterating the allowed types or using this regex.

Awaiting feedback then I'll either iterate the existing types or
refactor the the regex into something more maintainable (if that's ever
possible).